% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotnet-package.R
\docType{package}
\name{dotnet-package}
\alias{dotnet}
\alias{dotnet-package}
\title{dotnet: 'knitr' Engine for C# and F# Languages and Wrapper for 'dotnet' CLI}
\description{
Adds a 'dotnet' knitr engine for use in R Markdown with code chunks
  written in C# and F# languages. Requires the .NET Core SDK.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bearloga/dotnet}
  \item Report bugs at \url{https://github.com/bearloga/dotnet/issues}
}

}
\author{
\strong{Maintainer}: Mikhail Popov \email{mikhail@mpopov.com} (@bearloga on Twitter)

}
\keyword{internal}
